+++ tests/testthat/test-fetch-survey.R
@@ -1,5 +1,3 @@
-context("Download a survey from qualtRics and pull it into R using the fetch_survey() function")
-
 test_that("fetch_survey() returns survey with default params", {
 
   skip_on_cran()
@@ -32,7 +30,6 @@ test_that("fetch_survey() returns survey with custom params", {
   vcr::use_cassette("fetch_survey_custom", {
     x <-
       fetch_survey("SV_0pK7FIIGNNM0sNn",
-                   force_request = TRUE,
                    start_date = "2015-01-01",
                    end_date = "2022-06-02 18:40:53",
                    unanswer_recode = 999,
@@ -64,7 +61,6 @@ test_that("fetch_survey() excludes variable classes when requested", {
   vcr::use_cassette("fetch_survey_exclude", {
     x <-
       fetch_survey("SV_0pK7FIIGNNM0sNn",
-                   force_request = TRUE,
                    start_date = "2015-01-01",
                    end_date = "2022-06-02 18:40:53",
                    unanswer_recode = 999,
@@ -89,7 +85,6 @@ test_that("fetch_survey() returns survey with only one QID", {
   vcr::use_cassette("fetch_one_qid", {
     x <- fetch_survey(
       "SV_56icaa9YAafpAqx",
-      force_request = TRUE,
       limit = 15,
       include_questions = c("QID9"),
       breakout_sets = FALSE
@@ -119,31 +114,6 @@ test_that("fetch_survey() returns survey with only one QID", {
 })
 
 
-
-test_that("fetch_survey() reads a stored survey in temporary directory if exists", {
-  # Store RDS file
-  data <- "SUCCESS"
-  curr.wd <- getwd()
-  setwd(tempdir())
-  on.exit(setwd(curr.wd))
-  saveRDS(data, "surveyID.rds")
-  # Query with all options
-  expect_message(
-    qualtRics::fetch_survey("surveyID"),
-    "Loading saved prior download"
-  )
-})
-
-test_that("Save directory exists for fetch_survey()", {
-  expect_error(
-    qualtRics::fetch_survey(
-      "1234",
-      save_dir = "/users/jasper/desktop/idonotexist"
-    ),
-    "does not exist."
-  )
-})
-
 test_that("Limit cannot be less than one", {
   expect_error(
     qualtRics::fetch_survey("1234", limit = 0),
@@ -166,6 +136,18 @@ test_that("unanswer_recode is integer-ish", {
   )
 })
 
+test_that("correct error for deprecated args", {
+  skip_on_cran()
+
+  expect_snapshot(
+    fetch_survey("1234", force_request = TRUE),
+    error = TRUE
+  )
+  expect_snapshot(
+    fetch_survey("1234", save_dir = "~/Desktop"),
+    error = TRUE
+  )
+})
 
 # Restore the credentials for other tests:
 qualtrics_api_credentials(api_key = holder_API, base_url = holder_URL)